+++ R/function_Spectra_qualityMetrics.R
@@ -81,7 +81,7 @@ qualityMetrics <- function(object) {
             "intensity10xChange",
             "numberEmptyChrom",
             "medianIntensityRtIqr",
-            "xicFwhm",
+            "xicFwhm", "peakBoundary", "peakWidth", "peakBeta",
             "ticQuantileRtFraction", "areaUnderTic", "areaUnderTicRtQuantiles",
             "areaUnderTicMs1", "areaUnderTicMs2"
         )